requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --no-emit-index-url requirements.in
#
asgiref==3.7.2
    # via
    #   -r requirements.in
    #   django
    #   django-cors-headers
attrs==23.2.0
    # via
    #   -r requirements.in
    #   glom
    #   jsonschema
    #   referencing
    #   semgrep
boltons==21.0.0
    # via
    #   -r requirements.in
    #   face
    #   glom
    #   semgrep
bracex==2.4
    # via
    #   -r requirements.in
    #   wcmatch
build==1.0.3
    # via
    #   -r requirements.in
    #   pip-tools
certifi==2023.11.17
    # via requests
cfgv==3.4.0
    # via
    #   -r requirements.in
    #   pre-commit
charset-normalizer==3.3.2
    # via
    #   -r requirements.in
    #   requests
click==8.1.7
    # via
    #   -r requirements.in
    #   click-option-group
    #   pip-tools
    #   semgrep
click-option-group==0.5.6
    # via
    #   -r requirements.in
    #   semgrep
colorama==0.4.6
    # via
    #   -r requirements.in
    #   semgrep
coverage==7.3.3
    # via -r requirements.in
defusedxml==0.7.1
    # via
    #   -r requirements.in
    #   semgrep
distlib==0.3.8
    # via
    #   -r requirements.in
    #   virtualenv
django==4.2.8
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.3.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
emoji==2.9.0
    # via -r requirements.in
exceptiongroup==1.2.0
    # via
    #   -r requirements.in
    #   semgrep
face==22.0.0
    # via
    #   -r requirements.in
    #   glom
filelock==3.13.1
    # via
    #   -r requirements.in
    #   virtualenv
glom==22.1.0
    # via
    #   -r requirements.in
    #   semgrep
identify==2.5.33
    # via
    #   -r requirements.in
    #   pre-commit
idna==3.6
    # via
    #   -r requirements.in
    #   requests
importlib-metadata==7.0.1
    # via
    #   -r requirements.in
    #   build
jsonschema==4.20.0
    # via
    #   -r requirements.in
    #   semgrep
jsonschema-specifications==2023.12.1
    # via
    #   -r requirements.in
    #   jsonschema
markdown-it-py==3.0.0
    # via
    #   -r requirements.in
    #   rich
mdurl==0.1.2
    # via
    #   -r requirements.in
    #   markdown-it-py
nodeenv==1.8.0
    # via
    #   -r requirements.in
    #   pre-commit
packaging==23.2
    # via
    #   -r requirements.in
    #   build
    #   semgrep
peewee==3.17.0
    # via
    #   -r requirements.in
    #   semgrep
pip-licenses==4.3.3
    # via -r requirements.in
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==4.1.0
    # via
    #   -r requirements.in
    #   virtualenv
pre-commit==3.6.0
    # via -r requirements.in
prettytable==3.9.0
    # via
    #   -r requirements.in
    #   pip-licenses
pygments==2.17.2
    # via
    #   -r requirements.in
    #   rich
pyproject-hooks==1.0.0
    # via
    #   -r requirements.in
    #   build
pytz==2023.3.post1
    # via
    #   -r requirements.in
    #   djangorestframework
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   pre-commit
referencing==0.32.0
    # via
    #   -r requirements.in
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   -r requirements.in
    #   semgrep
rich==13.7.0
    # via
    #   -r requirements.in
    #   semgrep
rpds-py==0.16.2
    # via
    #   -r requirements.in
    #   jsonschema
    #   referencing
ruamel-yaml==0.17.40
    # via
    #   -r requirements.in
    #   semgrep
ruamel-yaml-clib==0.2.8
    # via
    #   -r requirements.in
    #   ruamel-yaml
semgrep==1.55.0
    # via -r requirements.in
sqlparse==0.4.4
    # via
    #   -r requirements.in
    #   django
tomli==2.0.1
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
    #   pyproject-hooks
    #   semgrep
typing-extensions==4.9.0
    # via
    #   -r requirements.in
    #   asgiref
    #   semgrep
urllib3==2.1.0
    # via
    #   -r requirements.in
    #   requests
    #   semgrep
virtualenv==20.25.0
    # via
    #   -r requirements.in
    #   pre-commit
wcmatch==8.5
    # via
    #   -r requirements.in
    #   semgrep
wcwidth==0.2.12
    # via
    #   -r requirements.in
    #   prettytable
wheel==0.42.0
    # via pip-tools
zipp==3.17.0
    # via
    #   -r requirements.in
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
